Bug 1198391: return summary information from state()
